net: samples: Modify zoap_server to new CoAP api's
