+++ samples/net/coap_server/Makefile
