+++ samples/net/coap_server/prj.conf
@@ -6,7 +6,7 @@ CONFIG_SYS_LOG=y
 CONFIG_SYS_LOG_SHOW_COLOR=y
 CONFIG_RANDOM_GENERATOR=y
 CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_ZOAP=y
+CONFIG_COAP=y
 CONFIG_NET_APP_SETTINGS=y
 CONFIG_NET_APP_MY_IPV6_ADDR="2001:db8::1"
 CONFIG_NET_IF_MCAST_IPV6_ADDR_COUNT=5
