+++ samples/net/coap_server/prj_bt.conf
@@ -17,4 +17,4 @@ CONFIG_SYS_LOG=y
 CONFIG_SYS_LOG_SHOW_COLOR=y
 CONFIG_RANDOM_GENERATOR=y
 CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_ZOAP=y
+CONFIG_COAP=y
