+++ samples/net/coap_server/prj_cc2520.conf
@@ -7,7 +7,7 @@ CONFIG_NET_LOG=y
 CONFIG_NET_UDP=y
 CONFIG_TEST_RANDOM_GENERATOR=y
 CONFIG_IEEE802154_CC2520=y
-CONFIG_ZOAP=y
+CONFIG_COAP=y
 CONFIG_NET_APP_SETTINGS=y
 CONFIG_NET_APP_IEEE802154_CHANNEL=20
 CONFIG_NET_APP_MY_IPV6_ADDR="2001:db8::1"
