+++ samples/net/coap_server/sample.yaml
