+++ samples/net/coap_server/src/Makefile
@@ -3,4 +3,4 @@ ccflags-y +=-I${ZEPHYR_BASE}/samples/net/common/
 ccflags-y +=-DNET_TESTING_SERVER=1
 endif
 
-obj-y = zoap-server.o
+obj-y = coap-server.o
